11.09.2025 Исправлены ошибки валидации данных, новые зависимости в requirements.txt

+++ embeddings_service/schemas.py
@@ -25,7 +25,7 @@ class HealthCheck(BaseModel):
     status: Literal["healthy", "failed"] = "healthy"
     model: str
     device: str = "cpu"
-    uptime: int | None = None
+    uptime: float | None = None
 
 
 class EmbeddingRequest(BaseModel):
